♻️ refactor: refactor the implement
