+++ packages/web-crawler/package.json
@@ -8,6 +8,7 @@
     "@mozilla/readability": "^0.5.0",
     "happy-dom": "^17.0.0",
     "node-html-markdown": "^1.3.0",
-    "query-string": "^9.1.1"
+    "query-string": "^9.1.1",
+    "url-join": "^5"
   }
 }
