+++ packages/web-crawler/src/urlRules.ts
@@ -61,4 +61,10 @@ export const crawUrlRules: CrawlUrlRule[] = [
     },
     urlPattern: 'https://www.qiumiwu.com/standings/(.*)',
   },
+
+  // mozilla use jina
+  {
+    impls: ['jina'],
+    urlPattern: 'https://developer.mozilla.org(.*)',
+  },
 ];
